Ad edit specific for user
Now a user can only edit their own post. There is no longer an edit button for every single ad. And if a user uses a direct URL to edit an ad that is not his the page will display that it is not his ad and display a button to go back